backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
scikit-learn==1.5.2
numpy==2.1.1
PyPDF2==3.0.1
python-docx==1.1.2
